py_server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
neo4j==5.28.1
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
sympy==1.14.0
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
